Fix PatientRepository generic type, persist prescriptions in PrescriptionController, add 17.png and 18.png screenshots, and include curl commands in outputs

+++ app/src/main/java/com/project/back_end/repo/PatientRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface PatientRepository extends CrudRepository<Patient, Integer> {
+public interface PatientRepository extends CrudRepository<Patient, Long> {
     Optional<Patient> findByEmail(String email);
 }
